requirements.txt
python-dotenv==1.0.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.115.8
h11==0.14.0
idna==3.10
motor==3.7.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
PyMySQL==1.1.1
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
